Moved CheckSystemAlarmScheduler functions to CheckSystemAlarm. Added SimpleNotify class to display display debugging information (will be removed later)
